database update

+++ Entities/Film.cs
@@ -16,7 +16,7 @@ namespace Entities
 		public string Description { get; set; }
 		public string Director { get; set; }
 		public string OriginCountry { get; set; }
-		public int Budget { get; set; }
+		public double Budget { get; set; }
 		[Column(TypeName = "float")]
 		public double RatingAvg { get; set; }
 		public IEnumerable<Comment> Comments { get; set; }
